centralize circuit solver
